+++ alerter/test/alerter/alerters/node/test_chainlink.py
@@ -66,7 +66,6 @@ class TestChainlinkNodeAlerter(unittest.TestCase):
         self.test_went_down_at_prometheus = None
         self.test_current_height = 50000000000
         self.test_total_block_headers_received = 454545040
-        self.test_no_of_active_jobs = 10
         self.test_max_pending_tx_delay = 6
         self.test_process_start_time_seconds = 345474.4
         self.test_total_gas_bumps = 11
@@ -85,7 +84,6 @@ class TestChainlinkNodeAlerter(unittest.TestCase):
         self.test_went_down_at_prometheus_new = None
         self.test_current_height_new = 50000000001
         self.test_total_block_headers_received_new = 454545041
-        self.test_no_of_active_jobs_new = 11
         self.test_max_pending_tx_delay_new = 7
         self.test_process_start_time_seconds_new = 345476.4
         self.test_total_gas_bumps_new = 13
@@ -184,10 +182,6 @@ class TestChainlinkNodeAlerter(unittest.TestCase):
                 'total_block_headers_received': {
                     'current': self.test_total_block_headers_received_new,
                     'previous': self.test_total_block_headers_received,
-                },
-                'no_of_active_jobs': {
-                    'current': self.test_no_of_active_jobs_new,
-                    'previous': self.test_no_of_active_jobs,
                 },
                 'max_pending_tx_delay': {
                     'current': self.test_max_pending_tx_delay_new,